Fixed test setup
