+++ backend/src/main/resources/application.yaml
@@ -1,5 +1,7 @@
 server:
   port: 8081
+  servlet:
+    context-path: "/api"
 spring:
   profiles:
     default: "dev"
