+++ backend/src/test/kotlin/com/beatgear/backend/annotation/IntegrationTest.kt
@@ -1,19 +1,11 @@
 import io.restassured.RestAssured
-import org.apache.http.client.utils.URIBuilder
-import org.junit.jupiter.api.BeforeAll
-import org.springframework.boot.json.JacksonJsonParser
 import org.springframework.boot.test.context.SpringBootTest
 import org.springframework.boot.test.util.TestPropertyValues
 import org.springframework.context.ApplicationContextInitializer
 import org.springframework.context.ConfigurableApplicationContext
-import org.springframework.http.MediaType
 import org.springframework.test.context.*
-import org.springframework.util.LinkedMultiValueMap
-import org.springframework.web.reactive.function.BodyInserters
-import org.springframework.web.reactive.function.client.WebClient
 import org.testcontainers.containers.PostgreSQLContainer
 import org.testcontainers.junit.jupiter.Testcontainers
-import java.util.*
 
 @Target(AnnotationTarget.CLASS)
 @Retention(AnnotationRetention.RUNTIME)
@@ -28,14 +20,7 @@ import java.util.*
 annotation class IntegrationTest {
     class Initializer : ApplicationContextInitializer<ConfigurableApplicationContext> {
 
-        lateinit var issuerUrl: String
-
         override fun initialize(applicationContext: ConfigurableApplicationContext) {
-            val environment = applicationContext.environment
-
-            issuerUrl = environment.getProperty("spring.security.oauth2.resourceserver.jwt.issuer-uri")
-                ?: throw IllegalStateException("Issuer URI is not configured!")
-
             TestPropertyValues.of(
                 "spring.datasource.url=${postgres.jdbcUrl}",
                 "spring.datasource.username=${postgres.username}",
